flask_app.py
from flask import Flask
import openai

# Your OpenAI API key
api_key = ''

# Set up the OpenAI client
openai.api_key = api_key

# Import the python-telegram-bot library
import telegram.ext

# Create a Flask app
app = Flask(__name__)

def generate_response(user_message):
  # Use the OpenAI API to send a request to ChatGPT
  response = openai.Completion.create(
      engine="text-davinci-002",
      prompt=user_message,
      max_tokens=1024,
      n=1,
      stop=None,
      temperature=0.5,
  )

  # Get the response from ChatGPT
  chatgpt_response = response['choices'][0]['text']

  return chatgpt_response

# Set up the bot with the Telegram API key
bot = telegram.ext.Updater(token='', use_context=True)

# Set up the bot to handle messages
@bot.message_handler(commands=['start', 'help'])
def send_welcome(update, context):
  context.bot.send_message(chat_id=update.effective_chat.id, text="Hi! I'm a ChatGPT bot. I can help you have a conversation with ChatGPT.")

@bot.message_handler(func=lambda message: True)
def handle_incoming_message(update, context):
  # Get the message from the user
  user_message = update.message.text

  # Use the ChatGPT handling code to generate a response
  response = generate_response(user_message)

  # Send the response back to the user
  context.bot.send_message(chat_id=update.effective_chat.id, text=response)

# Run the bot
bot.start_polling()
bot.idle()
