requirements.txt
flask==2.1.2
openai==0.26.0
python-telegram-bot==20.0
